Moved controls to snake.lua, deleted controls.lua
